+++ qunit/index.js
@@ -14,7 +14,7 @@ Inputmask.extendDefaults({
     inputEventOnly: false
 });
 
-// window.Inputmask = Inputmask; //inject globally for the simulator to detect inputeventonly
+window.Inputmask = Inputmask; //inject globally for the simulator to detect inputeventonly
 
 import tests_alternations from "./tests_alternations";
 import tests_attributes from "./tests_attributes";